likeSwitchメソッド作成

+++ pages/index.vue
@@ -7,9 +7,8 @@
         <div class="post-unit">
           <span class="post-author">{{ item.user.name }}</span>
           <span class="post-author">{{ item.like }}</span>
-          <img @click="like(item)" class="post-icon" src="@/assets/images/heart.png" alt="heart-icon">
-          <span class="">1</span>
-          <img @click="unlike(item)" class="post-icon" src="@/assets/images/heart.png" alt="heart-icon">
+          <img @click="likeSwitch(item)" class="post-icon" src="@/assets/images/heart.png" alt="heart-icon">
+          <span class=""></span>
           <img @click="deletePost(item.id)" class="post-icon" src="@/assets/images/cross.png" alt="cross-icon">
           <img @click="goDetail(item.id)" class="post-icon" src="@/assets/images/detail.png" alt="detail-icon">
         </div>
@@ -39,21 +38,30 @@ export default {
     goDetail(id) {
       this.$router.push('/post/' + id);
     },
+    likeSwitch(item) {
+      if (this.hasLiked(item)) {
+        this.unlike(item);
+      } else {
+        this.like(item);
+      }
+    },
     async like(item) {
-      await this.$axios.post("http://127.0.0.1:80/api/like/",
-        {
-          post_id: item.id,
-          user_id: item.user.id,
-        }
-      );
+      await this.$axios.post("http://127.0.0.1:80/api/like/", {
+        post_id: item.id,
+        user_id: item.user.id,
+      });
       this.getPost();
     },
     async unlike(item) {
       const like = item.like.find(like => like.user_id === item.user.id && like.post_id === item.id);
-      await this.$axios.delete("http://127.0.0.1:80/api/like/" + like.id);
-      this.getPost();
+      if (like) {
+        await this.$axios.delete("http://127.0.0.1:80/api/like/" + like.id);
+        this.getPost();
+      }
     },
-
+    hasLiked(item) {
+      return item.like.some(like => like.user_id === item.user.id && like.post_id === item.id);
+    },
   },
   created() {
     this.getPost();
